file renaming - predict api
