+++ src/predict/api/README.md
@@ -1,2 +1,2 @@
-#Predict API Guide
-Please refer to the main README.md for the predict microservice for more details on how to  and its components.
+# Predict API Guide
+Please refer to the main `README.md` for the predict microservice for more details on how to  and its components.
