+++ src/predict/api/dev-requirements.txt
@@ -1,0 +1,15 @@
+python-dotenv~=1.0.0
+Flask==2.3.2
+pymongo==4.4.1
+flask-cors==4.0.0
+flask-pymongo~=2.3.0
+flask-caching~=2.0.2
+redis==4.6.0
+gunicorn==21.2.0
+pandas
+google-cloud-bigquery==3.11.4
+Flask-SQLAlchemy
+psycopg2-binary
+GeoAlchemy2
+geojson
+pytest
